refactor(flow-log-upload): move HTTP ops into ingest module
The ingest connection, its requests and the redirect handling live in
their own module. The endpoint is kept as a parsed `Url` for the life of
the client instead of being re-parsed on every request and redirect, and
the client caps how many redirects it follows itself.
